App.css offloaded, refactored few components
Moved most of App.css styles to components' css, renamed and separated few components
