+++ src/components/Maps.css
@@ -1,0 +1,7 @@
+#center-pin {
+  position: absolute;
+  z-index: 1;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -100%);
+}
