+++ src/components/Maps.tsx
@@ -1,5 +1,6 @@
 import React, { useMemo } from 'react';
 import GoogleMapReact from 'google-map-react';
+import './Maps.css';
 
 const GOOGLE_API_KEY = process.env.REACT_APP_GOOGLE_API_KEY || '';
