+++ src/components/SplashScreen.css
@@ -1,0 +1,34 @@
+.splash {
+  text-align: center;
+}
+
+.splash-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .splash-logo {
+    animation: splash-logo-spin infinite 10s linear;
+  }
+}
+
+.splash-header {
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+@keyframes splash-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
